refactor(RandomMovement): 전진 조건 관련 구현 패키지 및 이름 수정
- 클래스의 의미를 조금 더 명확하게 하기 위함
