+++ src/main/java/racingcar/model/MovementCondition.java
@@ -1,0 +1,5 @@
+package racingcar.model;
+
+public interface MovementCondition {
+    int moving();
+}
